requirements.txt
numpy
pandas
openpyxl
mysql.connector-python
pymysql
python-dotenv
scikit-learn
##-e .